Added Monster Closet Elements (#143)
This commit contains 4 new models, and one Actor BP.

The Blueprint is just the compiled models, in their general proper positions. No Logic has been added yet.

The models are for the Door Top, Door Bottom, Wall, and damaged door, for when the spawn point has been 'destroyed'.

This should be enough to get the functional logic moving forward for these monster closets. We'll need to do FX and lots of other things eventually, but at least we can get the deployment working correctly.