Adding additional Aim Offsets and Cannam Poses and Anims
There are new Aim Offsets for Megaman, where he only moves his head.

There are new Aim Offsets for the Mirumijee, so they can target/look at Megaman when detected, or be used to aid in transitions as the enemy moves to attack the player.

There is a standing pose and hanging pose for the Cannam, along with a Standing walk animation to continue building from.