Added Rocks, Small Plants, Remaining Trees. Added Ambience.
Big update to the island. Lots more painted areas in addition to a ton more instanced foliage added. Currently too heavy for my computer. So it needs to be thinned up a lot.

We need player navigation figured out. Whether or not we want the player to fully explore the island or just portions, and then we can thin the forestation and speed things up a bit more. This also makes no use of level streaming. I'm confident this could really help a lot. Though we may need to generate several foliage elements instead of one big one for that to work properly.

Lastly, manual placement of static meshes and 'art directed' placement areas are still very much necessary. I'd like to focus on player navigation area before really addressing this though.