Added Refractor Box with 3 variants. Rusted, Dusty, and Pristine. Created new blueprints for each by duplicating the original containter blueprint. Added them sporadically throughout the level.
Noticed a number of issues when trying to make these boxes dynamic.

There is a static mesh component, which can house the actual static mesh, but instead the mesh is lower in the hierarchy, and there is also a box which i think is acting as a collision volume. Somehow, somewhere, the dynamic targeting is locked on to the location of the originally placed object, not the dynamic one.
So if you lock on to the dynamic box, it will fly around as you'd expect, except that your camera orbits the original location, while your lockon reticle widget follows the object. Very strange behavior.

In addition, apparantly the lock on doesn't like 3 duplications of the Container blueprint. It only wants to lock on to the dusty version for some reason. Would love to know the answers to these questions.