Mega Man Hang Anims and More... (#209)
* Added Refractor Box with 3 variants. Rusted, Dusty, and Pristine. Created new blueprints for each by duplicating the original containter blueprint. Added them sporadically throughout the level.

Noticed a number of issues when trying to make these boxes dynamic.

There is a static mesh component, which can house the actual static mesh, but instead the mesh is lower in the hierarchy, and there is also a box which i think is acting as a collision volume. Somehow, somewhere, the dynamic targeting is locked on to the location of the originally placed object, not the dynamic one.
So if you lock on to the dynamic box, it will fly around as you'd expect, except that your camera orbits the original location, while your lockon reticle widget follows the object. Very strange behavior.

In addition, apparantly the lock on doesn't like 3 duplications of the Container blueprint. It only wants to lock on to the dusty version for some reason. Would love to know the answers to these questions.

* Created duplicate PBR Base that has a noise flicker panner feature attached to the Emissive Channel. Then Built new Material instances off this new master base material. Timed and added the flicker so it feels mostly natural and part of the background. Set it up for all 3 Refractor Boxes.

* Added Treature Chest Shader and Texture Fixes. Added new Shield Generator Static Mesh with shaders and textures. Replaced their static meshes in their respective blueprints.

* added Hang Pullup, and Default Hang Pose. Need to add shuffle left and right, as well as hanging idle pose.

* Added New Megaman Hanging Idle Anim. Added Lightning Sparks FX to the Shield Generator. Added flicker to the lighting of the generator. Added lightning pack from marketplace (for the arcing of the generators)

* Added Hanging Shuffle anims, as well as a new Aim Offset for the special weapons arm.